MFT r20901: Do not assume that GLint is equivalent to long. Just use GLint.
git-svn-id: https://vrjuggler.googlecode.com/svn/juggler/branches/2.2@20902 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ modules/vrjuggler/vrj/Draw/OGL/GlViewCocoa.mm
@@ -80,7 +80,7 @@
          @throw ex;
       }
 
-      long renderer_id;
+      GLint renderer_id;
       [pixel_format getValues:&renderer_id
                  forAttribute:NSOpenGLPFARendererID
              forVirtualScreen:0];
